requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.1
django-autocomplete-light==3.11.0
django-environ==0.11.2
django-tempus-dominus==5.1.2.17
django-tinymce==4.0.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2023.4
